Patched TurnHandler class.
TurnHandler originally used the ITurnAgent interface - as much as I like abstractions, it unfortunately can't hold a reference to the actual deriving class which means if that character dies or their monobehaviour is disabled in some way; the TurnHandler will still try to call their behaviour.

That obviously causes a LOT of errors :)

Now characters can be deleted or removed entirely and the turnhandler will simply remove their reference from the turn agent pool.