Patched Inventory collection.
The DecreaseTotalSize method was removing the index of [^1] - meaning it was getting the last present item in the list and removing it.

If we had a collection size of 10, and wanted to decrease it by 2 ( to 8 total ) and we only had 5 items in the list currently, it would remove item 5 and 4 as it's removing the index of [^1] instead of removing the index of 9 and 8.