Fixed TurnHandler class - ITurnAgent is back.
TurnAgent seemed cumbersome for effectively just being inherited by deriving classes for the TurnAction() method.

Originally I went with and interface for an ITurnAgent - however in trying to delete the reference in the TurnHandler once the character had been deleted, it would throw an error.

For some reason the ITurnAgent == null doesn't work, but ITurnAgent.Equals( null ) works as intended. I'm sure it has to do with interfaces in general overriding the == operator.

I'm happy now, ITurnAgent is back y'all, it's BACK!