Patched GridMap class.
How this very simple error eluded me for so long is beyond me - when calculating the Origin of the GridMap my computation looked like this

CellCenter - new Vector2( Width * CellCenter.x, Height * CellCenter.y )

This is all well and good, it gets the origin and works perfect, right?

Yes and no - if a GridMap is a perfect square then this computation works great, it gets the bottom lefthand tile of the grid and computes from there. HOWEVER, if the GridMap is not square, or if one side is larger then the other, this computation will still calculate it as being effectively in the "middle" of the world at 0, 0.

Especially since I'm using Tilemaps for this which can be any size they want at startup, I had to fix it.

It's actually easier too, not the GridMap takes a startX and startY position on the constructor, those two values are basically our grid's origin - that is - the first position where everything else will calculate from.

Super simple fix, boggled my mind for a good hour, see ya later.