Added object pooling for EntityPawns and can display items on ground.
EntityPawns are basically just blank objects with a sprite renderer - I pooled these into an object pool and now whenever I need to display an item I can just grab it, display a sprite, and disable it when I'm done - it all gets all handled by the pool dynamically.

I currently set it up to the delegate event of the inventory and it seems to be working great so far. SpriteRenderers are cool because I can also tell if the EntityPawn is "visible" to the camera, meaning tomorrow or whenever I get around to it I can make the object cull when not on screen and get added back to the EntityPool - super efficient!

That's it for me tonight though, have a good one :)