Reworked entire asset and content pipeline.
First off, the Roguelike singleton I had before wasn't utilizing proper lazy instantiation, this meant if an object ( notably in the constructor of the object ) tried to call the Roguelike singleton's Instance reference before it was initialized as a monobehaviour, it would throw huge errors.

The way I fixed this was utilizing one of my favorite things in Unity - RequireComponent(). This allows me to add all of the core functionality once the Instance is called, and therefore allow me to use the AssetManager for constructor usage or the GameManager, etc...

Next to that, the old Addressable system worked great, but utlimately didn't allow me to do the above method of grabbing info and data before the monobehaviour had time to initalize fully - even through lazy instantiation.

I moved over to the StreamingAsset folder usage as this is outside of the projects internal file structure and can be called any time given you know how to read the data there-in.

I'm super super happy to say that not only do all of these systems work great now and can be called in constructors, but using the StreamingAssets folder allows players of the game to make mods! atleast content wise :)

Since the JSON file list is also in the StreamingAssets folder, players can make custom characters, throw in the proper texture for them, and have it loaded into the game normally, YAY!

What an absolute headache, but it was worth y'all.